changing all interfaces to represented as ints instead of bytes (bytes only go up to 127)
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@29035 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ firewall/impl/com/untangle/node/firewall/FirewallMatcher.java
@@ -113,7 +113,7 @@ public class FirewallMatcher
 
 
     public boolean isMatch( Protocol protocol,
-                            byte srcIntf, byte dstIntf,
+                            int srcIntf, int dstIntf,
                             InetAddress srcAddress, InetAddress dstAddress,
                             int srcPort, int dstPort)
     {
@@ -125,7 +125,7 @@ public class FirewallMatcher
 
     }
 
-    public boolean isMatchIntf(byte src, byte dst)
+    public boolean isMatchIntf(int src, int dst)
     {
         return this.srcIntf.isMatch(src) && this.dstIntf.isMatch(dst);
     }
